Add joypad and speed up rendering
The joypad receives press and release events from the event loop
which it can then use to update its own state. It's addressable so the
memory bus can refer to it at address 0xff00.

By caching decoded tile data in the ppu, rendering is sped up since
tile data is infrequently updated relative to how many times it is read.